Make defaultChild a getter, remove unused comments, fix lint violations, move captions settings menu data and methods to a utils file
